Add docker_up.sh, docker_down.sh, and services_ping.sh
Added Bash Scripts to better align with their PowerShell equivalents: docker_up.sh which is a bash version of docker_up.ps1 script, taking in -r and -b commands. docker_down.sh which is a bash version of docker_down.ps1.  services_ping.sh which is a bash version of services_ping.ps1.